+++ apps/spacerage/src/scenes/gameplay/Bullets.tsx
@@ -8,9 +8,7 @@ export const Bullets = () => (
     <planeGeometry args={[0.1, 0.8]} />
     <meshBasicMaterial color={new Color("yellow").multiplyScalar(2)} />
 
-    <ECS.ArchetypeEntities archetype="bullet">
-      {({ bullet }) => bullet}
-    </ECS.ArchetypeEntities>
+    <ECS.Archetype with="bullet">{({ bullet }) => bullet}</ECS.Archetype>
   </InstancedParticles>
 )
 
@@ -19,7 +17,7 @@ export const spawnBullet = (
   quaternion: Quaternion,
   velocity: Vector3
 ) =>
-  ECS.world.createEntity({
+  ECS.world.add({
     age: 0,
     destroyAfter: 1,
     velocity,