+++ apps/spacerage/src/scenes/gameplay/FollowCamera.tsx
@@ -1,4 +1,5 @@
 import { useFrame } from "@react-three/fiber"
+import { useEntities } from "miniplex/react"
 import { useEffect } from "react"
 import { useStore } from "statery"
 import { Vector3 } from "three"
@@ -11,7 +12,7 @@ const playerPos = new Vector3()
 const tmpVec3 = new Vector3()
 
 export const FollowCamera = () => {
-  const [player] = ECS.useArchetype("player", "sceneObject")
+  const [player] = useEntities(ECS.world.with("player", "sceneObject"))
 
   const { camera } = useStore(store)