+++ apps/spacerage/src/scenes/gameplay/GameplayScene.tsx
@@ -6,19 +6,16 @@ import { Vec3 } from "shader-composer"
 import { Color } from "three"
 import { Skybox } from "../../common/Skybox"
 import { Stage } from "../../configuration"
-import { MiniplexInspector } from "../../editor/MiniplexInspector"
 import { ScenePass } from "../../lib/ScenePass"
 import { SidebarTunnel } from "../../state"
 import { Nebula } from "../menu/vfx/Nebula"
 import { Asteroids } from "./Asteroids"
 import { Bullets } from "./Bullets"
-import { PlayerInspector } from "./editor/PlayerInspector"
 import { FollowCamera } from "./FollowCamera"
 import { HUD } from "./HUD"
 import { Pickups } from "./Pickups"
 import { Player } from "./Player"
 import { PostProcessing } from "./PostProcessing"
-import { ECS } from "./state"
 import { AgeSystem } from "./systems/AgeSystem"
 import { AttractorSystem } from "./systems/AttractorSystem"
 import { BulletSystem } from "./systems/BulletSystem"
@@ -82,10 +79,7 @@ const GameplayScene = () => {
           <AsteroidExplosions />
           <BackgroundAsteroids />
 
-          <SidebarTunnel.In>
-            <MiniplexInspector world={ECS.world} />
-            <PlayerInspector />
-          </SidebarTunnel.In>
+          <SidebarTunnel.In>{/* Extra inspectors */}</SidebarTunnel.In>
 
           <AgeSystem />
           <DestroyAfterSystem />