+++ apps/spacerage/src/scenes/gameplay/Pickups.tsx
@@ -1,27 +1,26 @@
 import { BallCollider, interactionGroups, RigidBody } from "@react-three/rapier"
+import { useEntities } from "miniplex/react"
 import { plusMinus } from "randomish"
 import { Color, Vector3 } from "three"
 import { InstancedParticles, Particle } from "vfx-composer-r3f"
-import { BECS, ECS, Layers, worldBucket } from "./state"
+import { ECS, Layers } from "./state"
 
 export const Pickups = () => (
   <InstancedParticles>
     <icosahedronGeometry args={[0.5]} />
     <meshStandardMaterial color={new Color("hotpink").multiplyScalar(4)} />
 
-    <BECS.Archetype properties="isPickup">
-      {({ entity }) => entity.jsx}
-    </BECS.Archetype>
+    <ECS.Archetype with="pickup">{(entity) => entity.pickup}</ECS.Archetype>
   </InstancedParticles>
 )
 
 export type PickupProps = { position: Vector3; onPickup?: () => void }
 
 export const Pickup = ({ onPickup, ...props }: PickupProps) => {
-  const [player] = ECS.useArchetype("player")
+  const [player] = useEntities(ECS.world.with("player"))
 
   return (
-    <BECS.Property name="rigidBody">
+    <ECS.Component name="rigidBody">
       <RigidBody
         {...props}
         linearVelocity={[plusMinus(4), plusMinus(4), 0]}
@@ -39,24 +38,23 @@ export const Pickup = ({ onPickup, ...props }: PickupProps) => {
           }
         }}
       >
-        <BECS.Property name="sceneObject">
+        <ECS.Component name="sceneObject">
           <BallCollider args={[0.5]}>
             <Particle />
           </BallCollider>
-        </BECS.Property>
+        </ECS.Component>
       </RigidBody>
-    </BECS.Property>
+    </ECS.Component>
   )
 }
 
 export const spawnPickup = (props: PickupProps) => {
-  const entity = worldBucket.add({
-    isPickup: true,
-    jsx: (
+  const entity = ECS.world.add({
+    pickup: (
       <Pickup
         {...props}
         onPickup={() => {
-          worldBucket.remove(entity)
+          ECS.world.remove(entity)
         }}
       />
     )