+++ apps/spacerage/src/scenes/gameplay/systems/DestroyAfterSystem.tsx
@@ -1,11 +1,10 @@
-import { archetype } from "bucketeer"
 import { Stage } from "../../../configuration"
 import { System } from "../../../lib/miniplex-systems-runner/System"
-import { ECS, worldBucket } from "../state"
+import { ECS, queue } from "../state"
 
-const entities = ECS.world.archetype("age", "destroyAfter")
-
-const withDestroyAfter = worldBucket.derive(archetype("destroyAfter", "age"))
+const entitiesToEliminate = ECS.world
+  .with("destroyAfter", "age")
+  .where((e) => e.age > e.destroyAfter)
 
 export const DestroyAfterSystem = () => (
   <System
@@ -13,18 +12,8 @@ export const DestroyAfterSystem = () => (
     world={ECS.world}
     updatePriority={Stage.Early}
     fun={() => {
-      for (const entity of entities) {
-        if (entity.age >= entity.destroyAfter) {
-          ECS.world.queue.destroyEntity(entity)
-        }
-      }
-
-      /* Squeeze Bucketeer in */
-      for (let i = withDestroyAfter.entities.length - 1; i >= 0; i--) {
-        const entity = withDestroyAfter.entities[i]
-        if (entity.age >= entity.destroyAfter) {
-          worldBucket.remove(entity)
-        }
+      for (const entity of entitiesToEliminate) {
+        queue(() => ECS.world.remove(entity))
       }
     }}
   />