+++ apps/spacerage/src/scenes/gameplay/systems/PlayerSystem.tsx
@@ -11,7 +11,7 @@ const tmpQuat = new Quaternion()
 const tmpVec3 = new Vector3()
 
 export const PlayerSystem = () => {
-  const entities = ECS.world.archetype("player", "rigidBody", "sceneObject")
+  const players = ECS.world.with("player", "rigidBody", "sceneObject")
   const fireCooldown = useRef(0)
 
   return (
@@ -21,7 +21,7 @@ export const PlayerSystem = () => {
         world={ECS.world}
         updatePriority={Stage.Early}
         fun={() => {
-          const [player] = entities
+          const [player] = players
           if (!player) return
 
           const { move, aim } = controller.controls
@@ -56,7 +56,7 @@ export const PlayerSystem = () => {
         world={ECS.world}
         updatePriority={Stage.Normal}
         fun={(dt) => {
-          const [player] = entities
+          const [player] = players
 
           if (!player) return
           const body = player.rigidBody.raw()