+++ apps/spacerage/src/scenes/gameplay/systems/VelocitySystem.tsx
@@ -2,7 +2,7 @@ import { Stage } from "../../../configuration"
 import { System } from "../../../lib/miniplex-systems-runner/System"
 import { ECS } from "../state"
 
-const entities = ECS.world.archetype("sceneObject", "velocity")
+const entities = ECS.world.with("sceneObject", "velocity")
 
 export const VelocitySystem = () => (
   <System