+++ packages/hmans-physics3d/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./colliders"
-export * from "./RigidBody"
-export * from "./types"
-export * from "./World"
-export * from "./interactionGroups"