+++ packages/shader-composer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # shader-composer
 
+## 0.4.8
+
+### Patch Changes
+
+- 08bb67b: **New:** Additional full name `Divide`, `Multiply`, and `Subtract` math helpers, aliased to `Div`, `Mul` and `Sub`, respectively.
+
+## 0.4.7
+
+### Patch Changes
+
+- 1557936: **New:** `$swizzle(value, swizzle)` returns a GLSL expression that applies swizzling to the given value.
+
 ## 0.4.6
 
 ### Patch Changes